+++ src/main/java/org/rumbledb/runtime/flwor/clauses/ForClauseSparkIterator.java
@@ -37,7 +37,7 @@ import org.rumbledb.runtime.flwor.closures.ForClauseSerializeClosure;
 import sparksoniq.jsoniq.ExecutionMode;
 import sparksoniq.jsoniq.tuple.FlworTuple;
 import sparksoniq.semantics.DynamicContext;
-import sparksoniq.spark.DataFrameUtils;
+import org.rumbledb.runtime.flwor.FlworDataFrameUtils;
 import sparksoniq.spark.SparkSessionManager;
 import sparksoniq.spark.udf.ForClauseUDF;
 
@@ -194,7 +194,7 @@ public class ForClauseSparkIterator extends RuntimeTupleIterator {
                 Dataset<Row> inputDF = this.child.getDataFrame(context, getProjection(parentProjection));
                 StructType inputSchema = inputDF.schema();
                 int duplicateVariableIndex = Arrays.asList(inputSchema.fieldNames()).indexOf(this.variableName);
-                List<String> columnsToSelect = DataFrameUtils.getColumnNames(
+                List<String> columnsToSelect = FlworDataFrameUtils.getColumnNames(
                     inputSchema,
                     duplicateVariableIndex,
                     null
@@ -205,7 +205,7 @@ public class ForClauseSparkIterator extends RuntimeTupleIterator {
                 } else {
                     columnsToSelect.add(expressionDFTableName + "`.`" + this.variableName);
                 }
-                String selectSQL = DataFrameUtils.getSQL(columnsToSelect, false);
+                String selectSQL = FlworDataFrameUtils.getSQL(columnsToSelect, false);
 
                 inputDF.createOrReplaceTempView(inputDFTableName);
                 expressionDF.createOrReplaceTempView(expressionDFTableName);
@@ -225,8 +225,8 @@ public class ForClauseSparkIterator extends RuntimeTupleIterator {
             Dataset<Row> df = this.child.getDataFrame(context, getProjection(parentProjection));
             StructType inputSchema = df.schema();
             int duplicateVariableIndex = Arrays.asList(inputSchema.fieldNames()).indexOf(this.variableName);
-            List<String> allColumns = DataFrameUtils.getColumnNames(inputSchema, duplicateVariableIndex, null);
-            Map<String, List<String>> UDFcolumnsByType = DataFrameUtils.getColumnNamesByType(
+            List<String> allColumns = FlworDataFrameUtils.getColumnNames(inputSchema, duplicateVariableIndex, null);
+            Map<String, List<String>> UDFcolumnsByType = FlworDataFrameUtils.getColumnNamesByType(
                 inputSchema,
                 -1,
                 this.dependencies
@@ -240,8 +240,8 @@ public class ForClauseSparkIterator extends RuntimeTupleIterator {
                     DataTypes.createArrayType(DataTypes.BinaryType)
                 );
 
-            String selectSQL = DataFrameUtils.getSQL(allColumns, true);
-            String UDFParameters = DataFrameUtils.getUDFParameters(UDFcolumnsByType);
+            String selectSQL = FlworDataFrameUtils.getSQL(allColumns, true);
+            String UDFParameters = FlworDataFrameUtils.getUDFParameters(UDFcolumnsByType);
 
             df.createOrReplaceTempView("input");
             df = df.sparkSession()
