+++ src/main/java/org/rumbledb/runtime/flwor/closures/ReturnFlatMapClosure.java
@@ -29,7 +29,7 @@ import org.rumbledb.api.Item;
 import org.rumbledb.runtime.RuntimeIterator;
 
 import sparksoniq.semantics.DynamicContext;
-import sparksoniq.spark.DataFrameUtils;
+import org.rumbledb.runtime.flwor.FlworDataFrameUtils;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -57,7 +57,7 @@ public class ReturnFlatMapClosure implements FlatMapFunction<Row, Item> {
 
         this.kryo = new Kryo();
         this.kryo.setReferences(false);
-        DataFrameUtils.registerKryoClassesKryo(this.kryo);
+        FlworDataFrameUtils.registerKryoClassesKryo(this.kryo);
         this.input = new Input();
     }
 
@@ -70,7 +70,7 @@ public class ReturnFlatMapClosure implements FlatMapFunction<Row, Item> {
         for (int columnIndex = 0; columnIndex < columnNames.length; columnIndex++) {
             String field = columnNames[columnIndex];
             if (dependencies.containsKey(field)) {
-                List<Item> i = DataFrameUtils.deserializeRowField(row, columnIndex, this.kryo, this.input); // rowColumns.get(columnIndex);
+                List<Item> i = FlworDataFrameUtils.deserializeRowField(row, columnIndex, this.kryo, this.input); // rowColumns.get(columnIndex);
                 if (dependencies.get(field).equals(DynamicContext.VariableDependency.COUNT)) {
                     this.context.addVariableCount(field, i.get(0));
                 } else {
@@ -97,7 +97,7 @@ public class ReturnFlatMapClosure implements FlatMapFunction<Row, Item> {
 
         this.kryo = new Kryo();
         this.kryo.setReferences(false);
-        DataFrameUtils.registerKryoClassesKryo(this.kryo);
+        FlworDataFrameUtils.registerKryoClassesKryo(this.kryo);
         this.input = new Input();
     }
 }
